mprove base structure of project
