+++ App.tsx
@@ -1,39 +1,12 @@
 import { StatusBar as ExpoStatusBar } from "expo-status-bar";
-import { StatusBar, StyleSheet, Text, View, SafeAreaView } from "react-native";
-import { Searchbar } from "react-native-paper";
-import React, { useState } from "react";
+import React from "react";
+import { RestaurantScreen } from "./src/feature/restaurants/screens/restaurants.screen";
 
 export default function App() {
-  const [searchQuery, setSearchQuery] = useState("");
   return (
-    <SafeAreaView style={styles.container}>
-      <View style={styles.search}>
-        <Searchbar
-          placeholder="Search"
-          onChangeText={setSearchQuery}
-          value={searchQuery}
-        />
-      </View>
-      <View style={styles.list}>
-        <Text>List</Text>
-        <ExpoStatusBar style="auto" />
-      </View>
-    </SafeAreaView>
+    <>
+      <RestaurantScreen />
+      <ExpoStatusBar style="auto" />
+    </>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: StatusBar.currentHeight,
-  },
-  search: {
-    padding: 20,
-    backgroundColor: "green",
-  },
-  list: {
-    padding: 20,
-    backgroundColor: "blue",
-    flex: 1,
-  },
-});
