+++ src/feature/restaurants/screens/restaurants.screen.tsx
@@ -1,0 +1,38 @@
+import { StatusBar as ExpoStatusBar } from "expo-status-bar";
+import { StatusBar, StyleSheet, Text, View, SafeAreaView } from "react-native";
+import { Searchbar } from "react-native-paper";
+import React, { useState } from "react";
+
+export const RestaurantScreen = () => {
+  const [searchQuery, setSearchQuery] = useState("");
+  return (
+    <SafeAreaView style={styles.container}>
+      <View style={styles.search}>
+        <Searchbar
+          placeholder="Search"
+          onChangeText={setSearchQuery}
+          value={searchQuery}
+        />
+      </View>
+      <View style={styles.list}>
+        <Text>List</Text>
+        <ExpoStatusBar style="auto" />
+      </View>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: StatusBar.currentHeight,
+  },
+  search: {
+    padding: 20,
+  },
+  list: {
+    padding: 20,
+    backgroundColor: "blue",
+    flex: 1,
+  },
+});
